+++ assets/js/navigation-mobile.js
@@ -14,7 +14,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.3.0
+ * @version  1.3.7
  */
 
 ( function() {
@@ -65,7 +65,7 @@
 		const
 			selectors = 'a, button, input:not([type=hidden]), select',
 			elements  = container.querySelectorAll( selectors ),
-			firstEl   = elements[0];
+			firstEl   = elements[0],
 			lastEl    = elements[ elements.length - 1 ],
 			activeEl  = document.activeElement,
 			tabKey    = ( 9 === event.keyCode ),