+++ templates/parts/admin/welcome-demo.php
@@ -7,7 +7,8 @@
  * @package    Michelle
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.12
+ * @since    1.0.12
+ * @version  1.3.7
  */
 
 // Exit if accessed directly.
@@ -19,42 +20,47 @@ if ( ! class_exists( 'WebManDesign\Michelle\Welcome\Component' ) ) {
 
 ?>
 
-<hr class="is-large">
+<div class="welcome__section welcome__section--demo" id="welcome-demo">
 
-<div class="about__section has-2-columns">
+	<h2>
+		<span class="welcome__icon dashicons dashicons-database-add"></span>
+		<?php esc_html_e( 'Theme Demo Content', 'michelle' ); ?>
+	</h2>
 
-	<h2 class="is-section-header"><?php esc_html_e( 'Theme demo content', 'michelle' ); ?></h2>
-
-	<div class="column">
+	<?php if ( get_option( 'fresh_site' ) ) : ?>
+	<div class="welcome__section--child">
 		<h3><?php esc_html_e( 'WordPress starter content', 'michelle' ); ?></h3>
 
 		<p>
 			<?php esc_html_e( 'This theme contains a predefined starter content.', 'michelle' ); ?>
-			<?php esc_html_e( 'So, if your WordPress installation is fresh and you haven\'t created any content yet, just open website customizer to preview the starter content.', 'michelle' ); ?>
-			<?php esc_html_e( 'Once you publish your customizer settings, the starter content will be ported into your website automatically.', 'michelle' ); ?>
+			<?php esc_html_e( 'So, if you haven\'t created any content yet, just open the customizer to preview the starter content.', 'michelle' ); ?>
+			<?php esc_html_e( 'Once you publish your customizer settings, the starter content will be imported into your website automatically.', 'michelle' ); ?>
 		</p>
 
-		<p><a href="<?php echo esc_url( admin_url( 'customize.php' ) ); ?>" class="button"><?php esc_html_e( 'Customize your website &raquo;', 'michelle' ); ?></a></p>
+		<p><em><small><?php esc_html_e( '(Note that WordPress starter content works only with fresh installation of WordPress.)', 'michelle' ); ?></small></em>		</p>
 
-		<p><em><small><?php esc_html_e( '(Note that WordPress starter content works only with fresh install of WordPress!)', 'michelle' ); ?></small></em>		</p>
+		<p><a class="button button-hero" href="<?php echo esc_url( admin_url( 'customize.php' ) ); ?>"><?php esc_html_e( 'Customize', 'michelle' ); ?></a></p>
 	</div>
+	<?php endif; ?>
 
-	<div class="column">
+	<div class="welcome__section--child">
 		<h3><?php esc_html_e( 'Full theme demo content', 'michelle' ); ?></h3>
 
 		<p>
-			<?php esc_html_e( 'Alternatively you can install a full theme demo content to match the theme demo website.', 'michelle' ); ?>
-			<?php esc_html_e( 'This provides a more complex data and a great fresh start to build your own website.', 'michelle' ); ?>
+			<?php esc_html_e( 'You can install a full theme demo content to match the theme demo website.', 'michelle' ); ?>
+			<a href="https://themedemos.webmandesign.eu/michelle/"><?php esc_html_e( '(Preview the demo &rarr;)', 'michelle' ); ?></a>
+			<?php esc_html_e( 'This provides a comprehensive start for building your own website.', 'michelle' ); ?>
 		</p>
 
 		<?php if ( class_exists( 'OCDI_Plugin' ) ) : ?>
-			<p><a href="<?php echo esc_url( 'themes.php?page=pt-one-click-demo-import' ); ?>" class="button"><?php esc_html_e( 'Install theme demo content &raquo;', 'michelle' ); ?></a></p>
-			<p><em><?php esc_html_e( 'Or head over to Appearance &raquo; Import Demo Data to start the import process.', 'michelle' ); ?></em></p>
+			<p>
+				<a class="button button-hero" href="<?php echo esc_url( 'themes.php?page=pt-one-click-demo-import' ); ?>"><?php esc_html_e( 'Install demo content', 'michelle' ); ?></a>
+				&ensp;
+				<small><em><?php esc_html_e( '(Appearance &rarr; Import Demo Data)', 'michelle' ); ?></em></small>
+			</p>
 		<?php else : ?>
-			<p><a href="https://webmandesign.github.io/docs/michelle/#demo-content" class="button"><?php esc_html_e( 'How to import demo content &raquo;', 'michelle' ); ?></a></p>
+			<p><a href="https://webmandesign.github.io/docs/michelle/#demo-content"><?php esc_html_e( 'How to import demo content &raquo;', 'michelle' ); ?></a></p>
 		<?php endif; ?>
-
-		<p><a href="https://themedemos.webmandesign.eu/michelle/"><?php esc_html_e( 'Preview theme demo &raquo;', 'michelle' ); ?></a></p>
 	</div>
 
 </div>