+++ templates/parts/admin/welcome-promo.php
@@ -8,7 +8,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.11
+ * @version  1.3.7
  */
 
 namespace WebManDesign\Michelle;
@@ -22,17 +22,19 @@ if ( ! class_exists( 'WebManDesign\Michelle\Welcome\Component' ) ) {
 
 ?>
 
-<hr class="is-large">
+<div class="welcome__section welcome__section--promo" id="welcome-promo">
 
-<div class="about__section is-feature welcome--promo">
 	<h2>
+		<span class="welcome__icon dashicons dashicons-superhero-alt"></span>
 		<?php esc_html_e( 'Like the theme?', 'michelle' ); ?>
 	</h2>
+
 	<p>
 		<?php esc_html_e( 'You are using a fully functional 100% free WordPress theme without any paid upgrade.', 'michelle' ); ?>
 		<?php esc_html_e( 'If you find it helpful, please support its updates and technical support service with a donation or by purchasing one of paid products at WebManDesign.eu.', 'michelle' ); ?>
 		<?php esc_html_e( 'Thank you!', 'michelle' ); ?>
-		<br><br>
-		<a href="https://www.webmandesign.eu/contact/#donation" class="button button-hero button-primary welcome--button-outline"><?php esc_html_e( 'Visit WebMan Design website now &rarr;', 'michelle' ); ?></a>
 	</p>
+
+	<p><a href="https://www.webmandesign.eu/contact/#donation"><strong><?php esc_html_e( 'Visit WebMan Design website now &rarr;', 'michelle' ); ?></strong></a></p>
+
 </div>